code/requirements.txt
numpy
torch
torchinfo
torchsummary
pyyaml